server/requirements.txt
# 服务端依赖
# 在服务端（有GPU的机器）安装

# 核心框架
langchain>=0.1.0
langgraph>=0.0.20
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM - ChatGLM3
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
cpm-kernels>=1.0.11

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.6.5
websockets>=12.0

# RAG
chromadb>=0.4.18
sentence-transformers>=2.2.2

# 数据处理
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
akshare>=1.11.0

# 技术指标
pandas-ta>=0.3.14b

# 工具
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1

# 日志监控
structlog>=23.2.0
prometheus-client>=0.19.0

# 缓存（可选）
redis>=5.0.1
